+++ packages/terra-core-docs/src/terra-dev-site/doc/hyperlink/example/AudioHyperlink.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import Hyperlink from 'terra-hyperlink';
 
-export default () => <Hyperlink href="https://www.cerner.com" variant="audio">Audio hyperlink</Hyperlink>;
+export default () => <Hyperlink href="https://www.cerner.com" variant="audio" text="Audio hyperlink" />;