+++ packages/terra-core-docs/src/terra-dev-site/test/hyperlink/UnderlineHiddenHyperlink.test.jsx
@@ -3,6 +3,6 @@ import Hyperlink from 'terra-hyperlink';
 
 export default () => (
   <div role="main">
-    <Hyperlink id="link" href="https://www.cerner.com" isUnderlineHidden>Underline hidden hyperlink</Hyperlink>
+    <Hyperlink id="link" href="https://www.cerner.com" isUnderlineHidden text="Underline hidden hyperlink" />
   </div>
 );