+++ packages/terra-core-docs/src/terra-dev-site/test/hyperlink/VideoHyperlink.test.jsx
@@ -3,6 +3,6 @@ import Hyperlink from 'terra-hyperlink';
 
 export default () => (
   <div role="main">
-    <Hyperlink id="link" href="https://www.cerner.com" variant="video">Video hyperlink</Hyperlink>
+    <Hyperlink id="link" href="https://www.cerner.com" variant="video" text="Video hyperlink" />
   </div>
 );