+++ packages/terra-hyperlink/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Changed
+  * updated hyperlink to have underline always.
+
+* Added 
+  * Added support for `hyperlink-button`.
+
 ## 2.57.0 - (February 15, 2023)
 
 * Changed